Src/backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
coolsms-python-sdk==2.0.3
Django==4.2.11
django-crontab==0.7.1
djangorestframework==3.15.1
djangorestframework-simplejwt==4.6.0
idna==3.7
PyJWT==2.8.0
requests==2.31.0
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
